chess-puzzle-challenger/requirements.txt
# Requirements for puzzle-challenger
python-chess>=1.9.0
pandas>=1.3.0
sqlalchemy>=1.4.0
click>=8.0.0
tqdm>=4.61.0
matplotlib>=3.4.0
pytest>=6.2.0
numpy>=1.20.0
